core_edge_v2/requirements.txt
hyperliquid-python-sdk>=0.9.0
streamlit>=1.35.0
pandas>=2.2.0
numpy>=1.26.0
plotly>=5.20.0
ta>=0.11.0
scipy>=1.13.0
optuna>=3.6.0
python-dotenv>=1.0.1
eth-account>=0.12.0
requests>=2.31.0
openpyxl>=3.1.2
xlsxwriter>=3.2.0
colorlog>=6.8.0
